+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - 🦺 Bugfix: Enabling Rich Presence on Discord after a game has already started would not show the game as being played. It now correctly updates within 5 seconds.
 - 🦺 Bugfix: Fixed a rare case where "Run on Startup" would be checked but would not actually start. This was due to the .exe being moved after "Run on Startup" was checked.
 - 🦺 Play Games Rich Presence will now only keep the current version's logs
+- ⚡️Added launch arg for hiding the tray icon on start
 
 ## v1.5.0